+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.61](https://github.com/react4j/react4j/tree/v0.61) (2018-04-06)
 [Full Changelog](https://github.com/react4j/react4j/compare/v0.60...v0.61)