+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ### Unreleased
 
 #### Changed
-* ð¥ **\[arez\]** Replace usage of the Arez `extra` and `browser-extra` spy utilities class with the separate
+* 💥 **\[arez\]** Upgrade Arez to version `0.76`.
+* 💥 **\[arez\]** Replace usage of the Arez `extra` and `browser-extra` spy utilities class with the separate
   top level project `arez-spytools`.
 
 ### [v0.61](https://github.com/react4j/react4j/tree/v0.61) (2018-04-06)